cloud/requirements.txt
Flask==2.3.2
gunicorn==21.2.0
playwright==1.37.0
numpy==1.24.3
pandas==2.0.3
asyncio
aiofiles==23.1.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
google-cloud-storage==2.10.0
tqdm==4.65.0